// src/models/convention.ts
export interface Convention {
    id: number;
    dateUpdate: string;
    isDimensionRoyale: boolean;
    idTypeConvention: number;
    typeConvention: string;
    txtColorTypeConvention: string;
    backgroundColorTypeConvention: string;
    idStatusConvention: number;
    statusConvention: string;
    txtColorStatusConvention: string;
    backgroundColorStatusConvention: string;
    description: string;
    dateSignature: string
}